+++ src/ViewRecord.js
@@ -1,6 +1,8 @@
 import _ from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
+import queryString from 'query-string';
+import { Redirect } from 'react-router';
 import Pane from '@folio/stripes-components/lib/Pane';
 import Icon from '@folio/stripes-components/lib/Icon';
 
@@ -29,13 +31,15 @@ function renderBody(record) {
 }
 
 
-// eslint-disable-next-line react/prefer-stateless-function
 class ViewRecord extends React.Component {
   static propTypes = {
     paneWidth: PropTypes.string.isRequired,
     onClose: PropTypes.func.isRequired,
     resources: PropTypes.shape({
       record: PropTypes.object,
+    }),
+    parentResources: PropTypes.shape({
+      query: PropTypes.object,
     }),
   }
 
@@ -50,6 +54,25 @@ class ViewRecord extends React.Component {
     const records = _.get(this.props.resources, ['record', 'records']) || [];
     const record = records[0];
 
+    if (record) {
+      const query = _.get(this.props.parentResources, 'query') || {};
+      let url =
+          (record.source === 'kb') ? `/eholdings/titles/${record.id}` :
+          (record.source === 'kb') ? `/inventory/${record.id}` :
+          undefined;
+      if (url) {
+        const obj = {};
+        if (query.qindex === 'title') {
+          obj.searchType = 'titles';
+          obj.q = query.query;
+        }
+        // eslint-disable-next-line prefer-template
+        url += '?' + queryString.stringify(obj);
+        return <Redirect to={url} />;
+      }
+    }
+
+    // Fallback: render the Codex record itself
     return (
       <Pane
         id="pane-recorddetails"